Create manga_scraper_services.py for the class, finalize main.py as the main API code to run
