--- backend_scraper/daily_scraper.py
@@ -1,1 +1,0 @@
-## This script is ran daily to webscrape the data to be displayed on the frontend
